+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # WIP
 
-- Lookup refs can be inlined in query :where patterns (issue #53)
+- Lookup refs can be inlined in query `:where` patterns (issue #53)
 - Pull API (issue #37, pull request #51) by [David Thomas Hume](https://github.com/dthume)
+- `pull` form in `:find` clause
 
 # 0.8.1